Rename toggle_user_membership as deactive_membership_by_user
It's not possible to know the state of the membership if each
call to the function toggles the state.

I could add an argument but this function is only used to disable
the membership.

At the first, I thought the user in argument was the user to toggle
so I change the function to clarify the used argument.