settings.test: Stop using logging.disable()
The function doesn't just prevent logs to be outputed; it prevent log
records to ever be emitted, which make impossible to test logging
behavior (with `TestCase.assertLogs` for exemple).

The original goal was to limit output cluttering when running the tests
suite so we can (almost) achieve the same thing with a custom LOGGING
configuration that send every log records to a `NullHandler`.
I said "almost" because if we add another handler to a logger somewhere
else then the log record will also be sent to that handler, but it's the
expected behavior of the `logging` module.