scripts.imports-asp: Automate further
If the previous files were not deleted `unzip` would failed, it expect
only one argument but the `*` was expanded by the shell, resulting in
two (or more) files given as arguments.
Also, our documentation tell us that we should delete the current week
files once the imports are done, but to keep a backup (just in case)
until the end of the week.

So I thought that we could change the script, so it works when there is
multiple files in the directory, and so we don't have to be bothered
with the backuping, then deleting, of the current week files.

And because there is a possibility that the files don't exists (ie.
script launched too early on monday) then we check they are here.

Also make shellcheck happy for the `cd` line.