Job Seeker creation & NIR update: fix "clone users"
The error happens for a yet unknown reason that we were unable to
successfully reproduce.

It all looks like some users have been created externally in the middle
of a user creation process, but that kind of race condition seems
extremely unlikely.

In the meantime, we'll just handle these errors more gracefully as no
straightforward explanation could be found. The session data is cleaned
as it should, the code behaves as expected, so "at first" it looks like
the issue IS rare and linked to a race condition.