__wip: Try and use CBVs to fix multiple issues.
- The "valid_session_required" decorator was only useful because the
session key was unfortunately the same whatever the SIAE we were filling
an application for. Since we will now use a different session key for
each SIAE, the decorator and its check becomes useless.

- A SessionDataMixin will come in handy to try and be more DRY with our
session management. An increasing number of our views are linked
together through the use of data shared within sessions, so let's try
and use this for now, so that we can generalize a little.

- The `back_url` parameter is also copy/pasted everywhere in the code,
handling it in a base view or mixin will help us to stop this
boilerplate.

Finally, all the views in `submit_views` rely on the presence of a Siae
instance in the session data. Let's materialize this through a base
class that enforces it.

Also, clean the session at the very end of the apply view series and
test that we can't get session info afterwards.

Bonus: rename submit_view to apply_views ??