notify_pole_emploi: Make a log more verbose
Right now a ton of logs are outputted that let us know that we did not
send a notification to PE, but we never know which approval and with
which dates.

It would be better in order to get a better understanding of this issue
to have comprehensive logs that do output the approval ID and dates.