settings.dev: Stop duplicating the logs of SQL queries
There already is a console handler configured for `django` in
`config.settings.base`, because logs propagate through the logger
hierarchy we don't need to duplicate the handler for `django.db.backends`
and only need to change the level to have the logs printed.

Also deleting the `require_debug_true` as it is unused, and we probably
will not use it anyway.