siaes.import_ea_eatt: Circumvent the missing emails problem
Until recently (start of may 2022) there were only a few rows with empty
emails, but in the last weeks we have always more than 20 missing emails
which lead our safe guard to stop the script prematurely.

The email is only required when we are creating a new SIAE, so move the
safe guard to a more appropriate place, one which will not block
everything because we are missing only some data.

Also add more informations in the output to be more self-explanatory.