www.apply: Ask all pertinents informations when creating a user
When an account was created by a proxy most of the job seeker's
informations were left out, but those informations are needed by the ASP
so they had to be filled by the SIAE when hiring the job seeker.
Because the prescriber usually have the job seeker at hand it make sense
to ask those informations at this moment, this should lighten the
load for SIAE when creating a "Fiche Salarié".