utils: Make the `SessionNamespace.NOT_SET` sentinel object falsy
The `NOT_SET` sentinel is returned by a `SessionNamespace()` when the
key doesn't exists so we can differentiate them from keys set to `None`.
Unfortunately Python's `object()` is truthy by default which could lead
to unexpected behavior when calling `.get()` in a `if` statement.
We could test the returned value against the sentinel object but it can
become quite cumbersome, so making it falsy allow us to keep a neat API
while keeping the possibility to differentiate between keys when needed.