users.models: Transform some `@property` into `@cached_property`
This is mainly to avoid duplicating SQL queries when accessing the
properties in views and templates.
This commit should be _relatively_ safe as the properties are related to
data that doesn't changes often, and if they changes it's in their
dedicated request. This also makes our requests more atomics by avoiding
possible race conditions.