config: Update default FQDN
We use 8080 since PR #126, because of PEAMU.
But we can now use PEAMU with the port 8000, so there's no need to use a
different port from django's default port.

Also : PEAMU does not allow localhost as base url for the callback uri,
but 127.0.0.1 is fine. Therefore I changed the dev FQDN to allow us to
use PEAMU without having to update the dev settings.