Use venv linters instead of fixed versions
There has been a mismatch in the versions from .pre-commit-config.yaml
and the versions in requirements. Instead of declaring the versions in
two places, use what’s available in the virtualenv.

That changes the expected developer set up, linters must be present in
the PATH (e.g. activated virtualenv). Before pre-commit would handle the
virtualenv on its own.

Discussing with the team, everybody set up a virtual environment
already, so that change should not be an issue.