staticfiles: Add "phantom" files to please Django 4.1
The newer ManifestStaticFilesStorage attempts to get the MAP files for
the resources it handles.

REST Framework doesn't yet provide the map files and at the time of
writing the latest release was 10 months old.

Maybe someday we'll be able to get rid of those, in the meantime let's
make collectstatic work.