settings: Move as many "constant" settings out of settings.base
And use them directly where needed. Settings should be variable.

In this commit we can't really show the disappearance of those settings
from the settings file as it would cause terrible conflicts.

But the commit still compiles and tests since the replacement is done in
the target files, only the cleanup isn't yet.