utils.apis: Split the huge test module in submodules
Originally, all the API tests were residing in a single, gigantic test
module that tests the entirety of the `itou.utils` app.

In order to keep the tests modules smaller and more to the point, and
also to make it easier to move the APIs later (in a dedicated app, for
instance), let's split those tests in more focused modules.