Use requirements/dev.in for dependencies in MacOS
The primary target of the app is Linux.
MacOS is available on a best-effort basis, but has a different set of
dependencies. For example, ipython pulls appnope (which is specific to
MacOS). Instead of maintaining a fixed set of requirements per platform,
pin and hash the dependencies for the main platform (Linux), and use a
looser specification for other platforms.