Install administrative criteria on post_migrate
When running the test suite twice (e.g. pytest && pytest), the second
execution always fails because the administrative criteria table is
empty.

That is caused by the presence of some TransactionTestCase subclasses in
the test suite, which call the flush management command on
`tearDownClass`.
https://github.com/django/django/blob/e9a24a15f2cdd6b4768ea7b45cd6b1758fb4ed48/django/test/testcases.py#L1304-L1312

Solved by installing the administrative criteria on the post_migrate
signal. That removes the need for some data migrations in
itou/eligibility/migrations.