Initialize the date pickers on every page
Previously, the date picker was served as a form asset (Media class).
The custom code used for initialization was meant to be executed only
once. Executing it multiple times caused an error (`const picker` was
defined multiple time). Example page where the failure could be
reproduced: the job application accept page.

Instead of using a form asset, load the script globally and run the
initialization once for the page.

To avoid polluting the global namespace, the global variables were
either moved to the function scope, or removed.