scripts: Add one-time commands to export the self-prescribers
This script has to be run at the beginning or end of any evaluation
campaign.

In 2021 we had a an SQL version which has now been lost and is restored
here for the sake of history.

In 2022 I'm introducing this Python-based version, untested but it's a
script that only has to be run once a year, with some variations.