Replace configure jobs preview jQuery with HTMX
Prefer the declarative style of HTMX.

Changes
=======

The view is contacted with a GET request instead of a POST. It’s only
reading data, it is safe.

The CSS used to disable them does not work for users with accessibility
needs, it’s only stylistic. The behavior should be consistent. Since
links in the preview direct users to a different page and don’t write on
the server, following them is not a big deal.

Tools
=====

The `data-htmx-open-modal` data attribute points to a modal in the DOM
that should be shown when the results of a request are available.