Implement evaluation campaign sanctions workflow
After an SIAE evaluation ends up with its proofs being refused, the
DDETS takes sanctions, which are announced through a notification. In
the evaluated SIAE list, where the system would previously display a
refused evaluation, it will now show a to_notify state.

When clicking the notify button, the DDETS user is directed to a page
where they input main reason for refusal, and detail the sanction. Once
saved, an email is sent to the SIAE members to notify them of the
result availability.

The SIAE can then log in and view the result of their evaluation.