Move EvaluatedSIAE to a final state on a closed evaluation campaign
When the campaign is closed, the state should always be either approved
or refused.

For in-progress evaluations, when the DDETS was waiting for the SIAE to
submit proofs, the evaluation is considered refused, since the SIAE
missed the deadline. When the SIAE submitted proofs, but the DDETS
failed to review them, the general rule in the administration is to
consider the proofs valid.

This forced transition may cause confusion, as an unsubmitted review
would still show. For example, the DDETS may have refused job
applications proofs, but not submitted their review. Once the campaign
is closed, the evaluation will be positive (because the DDETS did not
reply in time), but job applications may show as refused. An explanation
banner appears when the evaluation is forced to a final state.

Stop waiting for the campaign to be closed to reach the final state.
Instead, as soon as the `final_reviewed_at` is set, show the
evaluated_siae in read only mode. This behavior is akin to the “phase
contradictoire”, as soon as an SIAE proofs have been reviewed (and
at least one has been refused), the “phase contradictoire” starts.