Remove custom error pages
The 403 and 500 URLs were never used. The 404 case is puzzling:

From
https://www.clever-cloud.com/doc/deploy/application/python/python_apps/
which document the `http.json` format:

> error_pages: configure custom files for error pages

The "/404" directive is not a file. I expected CleverCloud to use the
`error_page` NGINX directive and that would cause :

> an internal redirect to the specified uri with the client request
  method changed to “GET”.

However, testing shows otherwise. All in all, static are not loaded by
users directly but by the main HTML page (statics are JavaScript and CSS
files, images, etc), so even a generic 404 is acceptable since regular
users will never see the HTML of that 404 page.