Spread antivirus scan tasks
The antivirus takes about 17 minutes to run for the daily scan. If a
deploy occurs in the middle of a scan, the scan is interrupted
(SIGTERM) and the files are never scanned. The full scan is expected to
take several days. No deploys in several days is quite unlikely, which
forces to track scan state and implement a resume mechanism.

S3 files are now frequently synced with the database by the
`sync_s3_files` command. With the long term goal of making Django the
source of truth for files, a files app has been added to store known
state.

Regularly, the `scan_s3_files` command checks the files table to find
files that needs scanning, takes up to 200 files, scan them and submits
the results. The batches were made smaller and the frequency of
processing increased, to facilitate recovering when a batch is
interrupted. This command can easily be scaled horizontally.

To follow signature evolutions, files are re-scanned monthly.

Miscellaneous changes:
- use absolute imports to follow the rest of the project
- read the bucket name from S3 settings