jobs: Remove old data files for ROMEs and Appellation
We will sync via API now. If we need those, they'll be in git.

Also, we renew the way those files were used during tests. There was a
test routine (used with EVERY SiaeFactory) that would load the entirety
of those gigantic JSON files, parse them and generate TENS OF THOUSANDS
of DB requests, in the worst cases.

Let's keep a single, short CSV file and load that one in TWO requests:
one for the ROMEs, one for the Appellations.

The repository grep will say thanks, and so will the speed of the tests.