jobs: Remove old commands to generate and import ROMEs
They are out of date since the scheme has changed and we won't
synchronize manually anymore.

If we ever need to get back at those, git is our friend. But they can't
work anymore, since we'll soon remove fields that were used with it and
the source JSON files are not to be found again in 2022.