jobs: Remove unnecessary fields
As Eric confirmed, none of those 3 fields that we kept on our models are
useful or ever will be.

Also, those are not easily available in the API (unless we ask for the
detail of every ROME object, which would be extremely inefficient with
around 530 HTTP requests)

So let's just get rid of those.