jobs: Add an updated_at column
We will be able to track changes on objects that might become severely
outdated.

Since we cannot realistically remove ROMEs and Appellation if they are
linked to Job Descriptions, at least we will be able to tell if those
are "old" or not have been updated in a long time.