siaes: Add source_id and source_kind to SiaeJobDescription
We will soon synchronise job offers from an external source, the first
of which will be the PE Api.

So we need to have a way to keep track of the original key of those
offers in the alien API to be able to do CRUD operations with every
synchronization.

Also keep track of the source, which for now can be either local or PE.