workflows: Remove import-asp, metabase-itou, final-tables.
import-asp is entirely unused, it is run by supportix and will soon be
automated.

metabase-itou is now run by a myriad of crons.

build-final-tables is ran at the end of every of our scripts, so at
the very least daily. Also, usage on github shows that manual runs are
not frequent, at most once every week but typically once a month.