populate_metabase_*: Remove the dry run mode.
This is not going to make Vermeer happy eventually, but I really think
it's the way to go.

build_metabase_itou is no more an experimental script, it's the
production basis for every data the C2 has access to. It must not break.

Same goes for build_final_tables and fluxiae.

The dry-run mode is confusing because it actually splits the script
into having two very different ways of functioning and did inject the
"dry-run" argument almost everywhere, like a tiny virus in the metabase
codebase. Every function had then 2 possible behaviours and outputs,
even though it would still write to the production database.

I recommend that all the testing for metabase should now be done with a
local postgres, or carefully in  a dedicated script, reusing what starts
to look like a very useful "metabase library" that we are building.

This way we will slowly ensure that only production ready stuff gets its
way to this command. I think it's the only way we will start to have a
cleaner code, and will find ways to still develop efficiently, but not
directly in production.

We also make `import_geiq` and `import_ea_eatt` use a `wet_run`
parameter instead of `dry_run`, so that in line with our other scripts,
the changes are actually only made when we explicitly ask for it.

Open to comments!