utils: Add a yield_sync_diff helper
It is used to display what changed between the newest data (coming from
a file, an API, ...) and a given queryset, whatever the model.

It is useful for Rome, Appellation and later on, Cities syncing as it
enables very useful info output that can be inspected to follow the
changes.