apis: Refactor the pole_emploi_api_client
Stop using a global client and switch to a lazy-loaded one.
This enables us to test it in a WAY easier fashion: we now only need to
mock the settings before the test, instead of monkey-patching the
global client from where it was imported.

Yes, even a global monkey-patch in itou.utils.apis would not work since
the object would already be resolved in memory. We could only patch use
by use.

Anyway !

In the same time, renovate some antiquated things:
- stop exposing internal URLs, this makes no sense and made the test
  reliant on internal details.
- use a shared cache for the API token, which is WAY better since now
  every thread using the client will benefit from this in the same time.
  Also, the code is simpler now and we don't need to store an ugly thing
  "in-memory", so less state.

We had to review a few tests.

Now everything is ready for use on a larger scale.

Downside: the client now relies on Django, for the cache.