SiaeJobDescription: Drop the (appellation, siae) unique constraint
It had been introduced a very long time ago, but since we want the SIAEs
to be able to publish job offers that could be in different places even
if the Appellation is the same.

On another note, the synchonisation with other sources (Pole Emploi for
instance) also forces us to drop this constraint since we don't actually
know or control the source SIAE for a given job offer.