metabase: Get rid of an unused constants.py
One of the constants should actually become a setting to be any useful.
The other constants are only used in this file so there's zero need for
another module.

Also, get rid of enable_sql_logging since the SQL_DEBUG env var in dev.py
does exactly that, does not log twice, and is only accessible in the
dev environment which is exactly what we want.