crons: Send slack messages as a django command, remove build_final_tables
This enables us to avoid spamming the C2 channel as much as possible.

Also, since we remove this, the build_metabase_final_tables was now a
single line, doing the same thing that populate_metabase_itou with the
right mode; get rid of it.

Also, reorganize a little the script executions and send slack
notifications accordingly.