settings: enable persistent connections
This should save ~ 15 ms on every requests according to Sentry.

Since the healthchecks are enabled, this should be quite safe:
instead of recreating a DB connection for every request, Django will
simply perform a "SELECT 1" request to check if the connection is re-usable
and create a new DB connection otherwise.