Inject fake scan data to spread the antivirus load
If all files are added to the database at once, the antivirus will start
scanning like crazy (its todo list just got 500k items). It will
complete its tasks as fast as possible, and be done in a few days. Next
month, it will be super busy for a few days, then do nothing for the
rest of the month.

Artificially inject scan data, so that the antivirus only scans a
portion of all files every day.