Improve email sending flow and testability
Always use the AsyncEmailBackend to send email, even in tests. That
backend spawns a Huey task that actually sends an email using backend
from the ASYNC_BACKEND_EMAIL setting.
In environments where asynchronous tasks are not available (i.e. not
production), Huey is configured in immediate mode, it will execute the
task instead of queuing it.

Instead of opening a connection for each email to be sent, use the same
connection and send all emails at once.

After this change, emails will use the same backend (AsyncEmailBackend)
and share more code paths between the tests and production settings,
leading to more realistic testing.

A bunch of documentation was outdated, erroneous or extraneous.
Performed a small sweep.