employee_record.management.commands: Add `clone_orphan_employee_records`
We don't try to be clever, we clone all orphans and let the SIAE decide
which one to fill (READY) and discard (DISABLED).

That should probably also be done in `move_siae_data` but I don't feel
comfortable to mess with employee records in that script as this can go
sideways really quick, and having some granularity can hurt.