siaes: Make source_kind nullable, unique with source_id, add source_url
This way, the "local" job descriptions would simply have None as their
source kind, and only the alien job descriptions would have this field
filled.

Also get rid of the blank=True since what we actually want are "real"
values but not blank strings, ever.

Finally, the main purpose of this whole work is to be able to provide an external
link to the offer, so we kind of should store it somehow ^^

Group everything in a single migration.