matomo: Run the sync at 4AM, Paris time
By experience, midnight seems to be a bad choice and Matromo tends stop
answering. I don't have the details of what's happening but it's
very possible that maintenance or consolidation operations happen around
that time.

I tried very successfully the same operation in the middle of the night,
so this mitigation seems to be a simple way to be more successful.