Implement cron monitoring
The crons were dead for a few days back in December, and it took a few
days to notice that. Use Sentry to monitor crons and be alerted earlier.

This first step only verifies one cron is running. There’s no API for
registering crons with Sentry, and registering them manually seems
tedious, and will likely result in an incomplete monitoring anyway.

Currently, missed crons result in synchronization delays, stale data and
analytics, but aren’t critical to the service. The recovery path is
trivial: run the cron. Therefore, this first step should do.