sync_pec_offers: Lock rows and use an atomic transaction
We had a Sentry issue where the synchronization did succed in raising a
database error, since we were creating objects that already were there.

This should not happen (even though it's not a big issue since they all
get recreated every hour) but it made me realise that:

- the rows that were inspected for creation, edition or deletion were
  not locked, so potentially subject to race conditions

- there was zero transaction around the multiple object modification
  that was taking place at the end of the command.

So far this is untested. It **could** be, but it's tedious.

Race condition: we have to simulate a race condition, cf how I did it
with the help of threads and black magic in:
https://github.com/betagouv/itou/blob/master/itou/approvals/tests/tests.py#L1500

Transaction: the most straightforward method would be monkey patching
Django's bulk_create and catch after running the command that the DB
state did not change; I find it a bit overkill as well.

Again, this data is overwritten every hour so...