autocomplete: Improve the search algorithm
The TrigramSimilarity that was in place is not great at finding
similarities between multiple words, but rather between single words.

The word trigram similarity is designed to find the greatest ordered set
of trigrams for a whole sentence and thus is better at finding matches
with French cities which are very frequently made of multiple words.

This really enhances our ability to find "Saint-Sulpice" and "Paris 8",
for instance (which TrigramSimilarity just does not match)

Also:
- only return resuklts where 50% of the trigrams do match.
- there was a swear words filter for an unknown reason.
- remove the 12 lines limit and setup a classic 10 results window.
- fine tune the sorting to further improve the results when they are
  scored the same.