populate_metabase_matomo: Turn a return into a continue
Matomo always only answers with a single CSV line for a given dashboard,
but by looking at the code it makes more sense to the reader to just
continue in the loop and avoid an extra exit condition to the calling
function. It's also more future proof if someday we use this function
for multiple results.

The tests do not change so far.