sync_pec_offers: Fix a PEBKAC issue seen on Sentry
Well, of course, and the test was opening a transaction so everything
was silent.

Now it should work as intended. Also remove the skip_locked in the
select_for_update as two concurrent processes would make the existing offers be
skipped in the second process and thus be created again.