itou.employee_record: Add the `ASPExchangeInformation()` abstract model
4 of the 5 fields were the same in both models: `asp_processing_code`,
  `asp_processing_label`, `asp_batch_file`, and `asp_batch_line_number`.
`archived_json` was only used in `EmployeeRecord()` but we have plan to
also fill it for `EmployeeRecordUpdateNotification()`

The `db_index=True` is removed for `asp_batch_file` and
`asp_batch_line_number` because it is more-or-less redundant with the
unique constraint (which is now created for both models).

Also, do some cleanup for `ASP_*` class attributes so they are all in the
new abstract model.