itou.users.models: Set `asp_uid` to `None` if it's not a job seeker
If the field is used in a form (like in the admin) then `""` instead of
`None` will be saved. Because there is an unique constraint on the field
only one `User` can use `""`, so we force the value to `None` if we are
not a job seeker. This will also handle job seeker users which have
their kind changed to another one.