Evaluation campaign: sanctions workflow
Once an evaluated SIAE has a negative outcome, the DDETS will pick
sanctions. After sanctions are chosen, the SIAE is notified and both the
DDETS and the SIAE can see the sanctions on their dashboard.

The notification email has been updated to be more precise, depending on
whether the SIAE is sanctioned.

In order to better handle the sanction date range, the
InclusiveDateRange helper has been introduced. It clearly states the
bounds are included and facilitates the handling of dates.