autocomplete: Improve city searches (order, accuracy, zip codes)
The search suffered from many issues.

It could return entries that had nothing to see with the query.
The order was seemingly random. The number of results was frustrating
when looking for a city that exists in many different departments.

Also there were issues related to the case and accent matching.

We could not lookup by post code either.

Address all of that.

Test a mode where we change the JS selector behaviour as well.