Simplify DuetDatePickerWidget
The `if value` is not necessary. NoneType and the empty string "" are
not instances of `datetime.date`.

The `isinstance(value, str)` block tries to parse the string as a date,
and returns the string. Should it fail to parse the string as a date ?
It will simply ignore the failure… There’s no point in doing work to
validate an input if we don’t look at the result.