employee_record: Fix cloning for conventions with the same ASP ID
`SiaeConvention()` are unique for the `asp_id`/`kind` pair so we can't
use `.get(asp_id=)` as this will lead to `MultipleObjectsReturned()`
being raised in some case.

Also improve the truthfulness and consistency of some factories by using
`SelfAttribute()` or existing traits.

Shorten SIAE's kind used in tests that are longer than 4 characters as
this is the maximum length of `Convention.kind`.