city_autocomplete: Fix a very long standing bug
A user called the support, saying that the search did not work on her
Chrome (mobile).

After investigation from François and I, it seems that:
- search is broken on any mobile device (as soon as you had touch
  emulation on)
- it has been so since a very long time

After investigation, it seems that the way it was implemented could not
work on mobile; it relied on using the `focus` event to set the city
name & slug appropriately in the form fields, but those were not fired
on mobile.

This new implementation aims to be simpler and rely on a more simple
chain of events, hopefully without regressions.