employee_record: Update denormalized fields on NEW and READY transition
It was done on SENT transition when the data was already sent, which
defeat the purpose to update the fields to take recent changes into
account, so we move it on the READY transition.

The NEW transition is used when reactivating an employee record, this is
mostly to check early that the new value will works, because of the
(`asp_id`, `approval_number`) unique constraint. Also, this will allow
to "refresh" an employee record.