Check user is authenticated before get_current_XXX_or_404
The LoginRequiredMixin does the authentication check in `dispatch()`,
which occurs after the `setup()` chain runs. That means the `setup()`
should deal with an AnonymousUser.

Many views requiring a session are protected with
`SessionNamespaceRequiredMixin` and we show a 403 because the expected
session key is missing, so the `AnonymousUser` is handled.

Some however, rely on custom `get_current_XXX_or_404()`, which check
properties and use methods on the custom User model, that AnonymousUser
does not provide.

Handle anonymous users gracefully by redirecting them to the login page.