Simplify EvaluatedSiae.state
Extract the condition that sets the evaluated SIAE state to accepted
when it has submitted documents and the DDETS did not answer, instead of
repeatedly checking it under different branches.

Enforce that `reviewed_at` must be set during a campaign (to simplify
the conditionals).
At the start of the adversarial phase (defined in the DGEFP planning),
all SIAEs that were not reviewed have their reviewed_at field set (by
calling transition_to_adversarial_phase()). It is safe to assume that
field is set.

Some tests created an invalid situation, where the campaign was closed
without the reviewed_at field being set. This has been discovered and
fixed.

`test_review` was not verifying anything meaningful (calling review() on
a closed campaign) and has been removed.

Co-authored-by: Xavier Fernandez <xavier.fernandez@beta.gouv.fr>