employee_record.clone_orphan_employee_records: Handle user created SIAE
SIAE created by a user (called "antenna" in the codebase) are sent with
the SIRET of the structure (the "mother") that is known by the ASP, they
also share the same ASP ID (used in a UNIQUE CONSTRAINT).
When an antenna become known by the ASP, SIRET and/or ASP ID can
changes, it become it's own "mother" but because the `siret` and
`asp_id` field are denormalized the employee record become "orphans" and
we decided to clone them to keep the history.

The problem to use `--old-asp-id` to filter is that in the case of
antennas we are going to take *all* the employee records and not just
the one of the SIAE.