matomo: fix a nasty bug that can prevent any write
Sometimes the order of execution makes it possible that we don't write
any of the results to metabase.

Explanation: we run 20 dashboard fetches in parallel.

Let's imagine that the 19 first to return get us valid values, but
unfortunately the 20th and latest is empty.

We would not detect correctly its emptiness and would make `columns`
be [], which would in turn make the subsequent check before updating
Metabase fail.

Not only because of this but because the Matomo table structure did
change, we need to re-run this from the beginning of 2022.

This is tricky to test though cause the order of the ThreadPoolExecutor
is by nature non deterministic so it is difficult to design mocks so
that the first URLs are filled and "the last one consulted" is empty.