Show number of weeks since job applications
Before, a blanket message would show “more than 3 weeks”. Now, after 3
weeks, the actual number of weeks is displayed.

The with_is_pending_for_too_long queryset method wasn’t widely used, and
was hard to update to the desired logic. Instead, it was moved to a
helper in the view.