Revert "requirements: Add pytest-timeout"
This reverts commit 2450bce82e35b6e38d07c87ed21f1c0a9db5bd9a.

The timeout includes the setup and teardown times. That’s what makes the
timeout the most useful, because e.g. a setup of fixtures can be
particularly slow and should be reported. Creating a test database and
populating it is considered part of the test setup (--create-db).

Some dev machines are particularly slow (e.g. a Mac M1 using docker with
amd64 emulation), and it takes more than 45 seconds for the first test
to start. Increasing the timeout to 60 seconds (lowest value for which
tests run reliably on said slow machine) loses most of the benefit of a
timeout.

https://github.com/betagouv/itou/pull/2346