address: Add geocoding_updated_at & ban_api_resolved_address
This will become very useful to run our geocoding updating scripts
regularly and apply various policies to them.

Furthermore, let's add a field that will be very useful to store the
resolved address (as returned by the BAN) to be able later to run manual
inspections or disambiguations between the user's address and the
API-returned one.