jobseeker_profile: Always have one, even in tests.
The current code still allows for passing a NULL JobSeekerProfile with
out JobSeekerFactory through jobseeker_profile=False, but it should not
be used or encouraged except for VERY SPECIFIC reasons.

Right now it is only kept for backwrds compatibility with all the
Employee Records tests who rely on a specific
JobSeekerProfileWithHexaAddress object to be created and available at
all times.